parallel-package {parallel}	R Documentation

Support for Parallel Computation

Description

Support for parallel computation, including random-number generation.

Details

This package was first included with R 2.14.0 in 2011.

There is support for multiple RNG streams with the "L'Ecuyer-CMRG" RNG: see nextRNGStream.

It contains functionality derived from and pretty much equivalent to that contained in packages multicore (formerly on CRAN, with some low-level functions renamed and not exported) and snow (for socket clusters only, but MPI and NWS clusters generated by snow are also supported). There have been many enhancements and bug fixes since 2011.

This package also provides makeForkCluster to create socket clusters by forking (not Windows).

For a complete list of exported functions, use library(help = "parallel").

Author(s)

Brian Ripley, Luke Tierney and Simon Urbanek

Maintainer: R Core Team R-core@r-project.org

See Also

Parallel computation involves launching worker processes: functions psnice and pskill in package tools provide means to manage such processes.

[Package parallel version 3.6.1 Index]